pyproject.toml
[project]
name = "repo-sync"
version = "0.1.1"
description = "Sync local git repositories with GitHub remotes"
requires-python = ">=3.12"
dependencies = ["discord-notify", "pyyaml>=6.0"]

[tool.uv.sources]
discord-notify = { git = "https://github.com/Seika139/discord-notify.git", tag = "v0.1.1" }

[project.scripts]
repo-sync = "repo_sync.cli:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.ruff]
target-version = "py312"
line-length = 100

[tool.ruff.lint]
select = ["E", "F", "I", "W", "UP"]

[tool.pytest.ini_options]
testpaths = ["tests"]

[tool.mypy]
python_version = "3.14"
strict = true
exclude = ["^build/", "^\\.venv/"]

[tool.rumdl]
respect_gitignore = true
exclude = [".claude/**"]
disable = [
  "MD013", # 行の長さは自由
  "MD024", # 重複した見出しを許可
  "MD025", # 複数のトップレベル見出しを許可
  "MD033", # HTML タグを許可
  "MD036", # 強調表示の過剰使用を許可
  "MD041", # ファイルの最初の行を見出しにする必要なし
]

[tool.rumdl.MD009]
strict = true

[dependency-groups]
dev = [
  "mypy>=1.18",
  "pytest>=8.0",
  "ruff>=0.11",
  "ty>=0.0.14",
  "types-pyyaml>=6.0.12.20260408",
]
